Moved the sineTable definition into a cpp file instead of having it in a header file.
You really should never implement a single variable like this in a header file, but I did it because I had trouble linking the game when the array was a const. Now I have figured out how to solve that (by declaring the variable as "extern C") so now I can fix sinetable.h. :)