Changed the name of the setup and loop functions so they don't collide with the functions in the sketch.
The plan right now to get the code to be cross platform is that the OSX entry point call tage_setup and
tage_loop, defined in the project. The project should then be set up as a library used by an Arduino
sketch. So the Arduino sketch that runs a game will sinply include the engine and the project, and then
call tage_setup() from setup() and tage_loop() from loop(). So the only thing differentiating between
games in the Arduino sketch will be which game library is included.