Defined audio source for title screen music.
It is a regular AudioSource now but the audio samples for this short song is way too many to fit
in Flash Memory of the TinyArcade. In fact, just the song is 11 KB bigger than what the entire
can use of Flash Memory. So I will probably either create something like a streamed audio
source, or write a simple audio synthezier.